requirements.txt
asgiref==3.2.10
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1
psycopg2-binary>=2.8
requests>=2.24
Pillow>=7.2.0
django-cors-headers>=3.5.0
django-user-unique-email>=0.1.1
drf-writable-nested>=0.6.1
